refactor(PieceType): 피스타입 기물이 반환
