+++ src/main/java/janggi/domain/piece/Cannon.java
@@ -1,6 +1,7 @@
 package janggi.domain.piece;
 
 
+import static janggi.domain.piece.PieceType.CANNON;
 import static janggi.domain.piece.direction.Direction.DOWN;
 import static janggi.domain.piece.direction.Direction.LEFT;
 import static janggi.domain.piece.direction.Direction.LEFT_DOWN;
@@ -126,5 +127,10 @@ public class Cannon extends Piece {
     @Override
     public double getScore() {
         return CANNON_SCORE;
+    }
+
+    @Override
+    public PieceType getPieceType() {
+        return CANNON;
     }
 }
